requirements.txt
backcall==0.1.0
cycler==0.10.0
decorator==4.4.0
Django==2.2.3
django-extensions==2.2.1
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.1.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
